// frontend/src/app/api/model/staffShipmentViewDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StaffShipmentViewDto { 
    id?: string;
    trackingNumber?: string;
    type?: StaffShipmentViewDto.TypeEnum;
    status?: StaffShipmentViewDto.StatusEnum;
    weight?: number;
    length?: number;
    width?: number;
    height?: number;
    totalPrice?: number;
    paidBy?: StaffShipmentViewDto.PaidByEnum;
    isPaid?: boolean;
    createdAt?: string;
    updatedAt?: string;
    senderId?: string;
    senderName?: string;
    senderPhone?: string;
    receiverId?: string;
    receiverName?: string;
    receiverPhone?: string;
    originOfficeId?: number;
    originOfficeName?: string;
    originAddressString?: string;
    deliveryOfficeId?: number;
    deliveryOfficeName?: string;
    deliveryAddressString?: string;
    currentOfficeId?: number;
    currentOfficeName?: string;
    currentCourierId?: string;
    currentCourierName?: string;
    registeredById?: string;
    registeredByName?: string;
    appliedAddons?: Array<string>;
}
export namespace StaffShipmentViewDto {
    export const TypeEnum = {
        Document: 'DOCUMENT',
        Parcel: 'PARCEL',
        Pallet: 'PALLET',
        Oversized: 'OVERSIZED'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const StatusEnum = {
        Registered: 'REGISTERED',
        InTransit: 'IN_TRANSIT',
        AtDeliveryOffice: 'AT_DELIVERY_OFFICE',
        OutForDelivery: 'OUT_FOR_DELIVERY',
        Delivered: 'DELIVERED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const PaidByEnum = {
        Sender: 'SENDER',
        Receiver: 'RECEIVER'
    } as const;
    export type PaidByEnum = typeof PaidByEnum[keyof typeof PaidByEnum];
}